refactor proto changes
