+++ src/main/java/org/dependencytrack/event/kafka/KafkaTopics.java
@@ -7,9 +7,9 @@ import org.cyclonedx.proto.v1_4.Bom;
 import org.dependencytrack.common.ConfigKey;
 import org.dependencytrack.event.kafka.serialization.KafkaProtobufSerde;
 import org.hyades.proto.notification.v1.Notification;
+import org.hyades.proto.repointegrityanalysis.v1.IntegrityResult;
 import org.hyades.proto.repometaanalysis.v1.AnalysisCommand;
 import org.hyades.proto.repometaanalysis.v1.AnalysisResult;
-import org.hyades.proto.repometaanalysis.v1.IntegrityResult;
 import org.hyades.proto.vulnanalysis.v1.ScanCommand;
 import org.hyades.proto.vulnanalysis.v1.ScanKey;
 import org.hyades.proto.vulnanalysis.v1.ScanResult;
