+++ src/main/proto/org/hyades/repometaanalysis/v1/repo_meta_analysis.proto
@@ -37,28 +37,4 @@ message Component {
   optional string sha1_hash = 4;
   optional string sha256_hash = 5;
   optional string uuid = 6;
-  optional double component_id = 7;
-}
-
-message IntegrityResult {
-  // The Component for which the integrity check was done
-  Component component = 1;
-  //md5 hash match status of component
-  HashMatchStatus md5_hash_match = 2;
-  //sha1 hash match status of component
-  HashMatchStatus sha1_hash_match = 3;
-  //sha256 hash match status of component
-  HashMatchStatus sha256_hash_match = 4;
-  //Identifier of the repository with which the hashes were compared
-  string repository = 5;
-  // When was the integrity check done
-  google.protobuf.Timestamp published = 6;
-}
-
-enum HashMatchStatus {
-  HASH_MATCH_STATUS_UNSPECIFIED = 0;
-  HASH_MATCH_STATUS_PASS = 1;
-  HASH_MATCH_STATUS_FAIL = 2;
-  HASH_MATCH_STATUS_UNKNOWN = 3;
-  HASH_MATCH_STATUS_COMPONENT_MISSING_HASH = 4;
 }
